Implements Solution
As noted in the comments, I am basically ignoring the attributes on the Serializer classes.
They are a roadmap to getting the the proper output, but I cannot figure out how to get
the struct data, which exists in an instance to do something useful with what look to me to be
method calls on the class.